+++ scripts/gulpfile.ts
@@ -3,9 +3,15 @@ import chpwd from "./change-pwd";
 import generate from "./generate";
 import genImgMap from "./generate-img-map";
 import subImg from "./substitute-img";
+import downloadImg from "./download-img";
 
 gulp.task("generate-image-map", async (cb) => {
   await genImgMap();
+  cb();
+});
+
+gulp.task("download-image", async (cb) => {
+  await downloadImg();
   cb();
 });