+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix `npm run chpwd` (change password) not working for block-level.
 * Add `npm run genimg` for generating img map.
 * Add `npm run subimg` for substituting img.
+* Add `npm run downimg` for downloading img.
 
 ## 2023.03.23
 * Folder refactoring